refactor: frame panels
